data_management: move dmu code
